fix(onboarding): support GPT-5 ping shape
